+++ src/Adapter/Status/STATUS.mjs
@@ -1,0 +1,21 @@
+export const STATUS_200 = 200;
+
+export const STATUS_201 = 201;
+
+export const STATUS_302 = 302;
+
+export const STATUS_400 = 400;
+
+export const STATUS_401 = 401;
+
+export const STATUS_403 = 403;
+
+export const STATUS_404 = 404;
+
+export const STATUS_405 = 405;
+
+export const STATUS_500 = 500;
+
+/**
+ * @typedef {STATUS_200 | STATUS_201 | STATUS_302 | STATUS_400 | STATUS_401 | STATUS_403 | STATUS_404 | STATUS_405 | STATUS_500} STATUS
+ */