+++ CHANGELOG.md
@@ -6,6 +6,15 @@ Changes:
 
 \-
 
+## v2023-01-03-1
+
+Changes:
+
+- Rename to `flux-http-api`
+- `fetch`
+- Constants
+- Reduce `express`
+
 ## v2022-12-08-1
 
 Changes: